Refactored grpc server to use both servicers
